fix: Fix Sqlite and other database dispatch
Database explorer dispatch was not working because dispatch was called
on the base class. To be able to cleanly use derived class dispatch,
Sqlite and other databases had to to be separated. So sqlite is a top
level option instead of being clubbed under 'db'

Sqlite tests were enabled and dispatch tests were fixed to reproduce the
bug.

Fix #36